Moving logging to its own package, so we can add a Configure() func to it.
